+++ src/teacher/teacher.service.ts
@@ -1,4 +1,4 @@
-import { ConflictException, ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
+import { ConflictException, ForbiddenException, Injectable, NotFoundException, UnauthorizedException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Teacher } from './entity/teacher.entity';
@@ -7,6 +7,8 @@ import { Status } from '../enums/exam.enum';
 import { Exam } from '../exam/entity/exam.entity';
 import { Room } from '../room/entity/room.entity';
 import { RoomStatus } from '../enums/room.enum';
+import { Token } from '../auth/token.decorator';
+import { Role } from '../enums/user.enum';
 
 @Injectable()
 export class TeacherService {
@@ -33,14 +35,32 @@ export class TeacherService {
         return teacher;
     }
 
-    async update(id: string, updateTeacherDto: UpdateTeacherDto): Promise<Teacher> {
+    async update(id: string, updateTeacherDto: UpdateTeacherDto, token: any): Promise<Teacher> {
+
+        const loggedInTeacher = await this.teacherRepository.findOne({
+            where: { user: { userId: token.id } },
+            relations: ['user'],
+        });
+
+        if (!loggedInTeacher) {
+            throw new UnauthorizedException('Logged-in user is not associated with a teacher');
+        }
+
+        if (loggedInTeacher.teacherId !== id) {
+            throw new ForbiddenException(`You are not authorized to update this teacher's information`);
+        }
+
         const teacher = await this.findOne(id);
 
         Object.assign(teacher, updateTeacherDto);
         return this.teacherRepository.save(teacher);
     }
 
-    async delete(id: string): Promise<void> {
+    async delete(id: string, token: any): Promise<void> {
+        if (token.role !== Role.ADMIN) {
+            throw new UnauthorizedException('Only admins can delete');
+        }
+
         const result = await this.teacherRepository.delete(id);
         if (result.affected === 0) {
             throw new NotFoundException(`Teacher with ID ${id} not found`);
@@ -55,7 +75,10 @@ export class TeacherService {
         return teacher;
     }
 
-    async updateRoomForExam(examId: string, roomIds: string[]): Promise<Exam> {
+    async updateRoomForExam(examId: string, roomIds: string[], @Token() token: any): Promise<Exam> {
+        if (token.role !== Role.TEACHER) {
+            throw new UnauthorizedException('Only teachers can update exam rooms');
+        }
         const exam = await this.examRepository.findOne({
             where: { examId },
             relations: ['teacher', 'rooms'],
